+++ AstarteDeviceSDKCSharp/Transport/MQTT/MutualSSLAuthenticationMqttConnectionInfo.cs
@@ -18,7 +18,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-using MQTTnet.Client.Options;
+using MQTTnet.Client;
 
 namespace AstarteDeviceSDKCSharp.Transport.MQTT
 {
@@ -26,11 +26,12 @@ namespace AstarteDeviceSDKCSharp.Transport.MQTT
     {
 
         private readonly Uri _brokerUrl;
-        private readonly IMqttClientOptions _mqttConnectOptions;
+        private readonly MqttClientOptions _mqttConnectOptions;
         private readonly string _clientId = string.Empty;
+        private readonly TimeSpan _timeOut;
 
         public MutualSSLAuthenticationMqttConnectionInfo(Uri brokerUrl, string astarteRealm,
-        string deviceId, MqttClientOptionsBuilderTlsParameters tlsOptions)
+        string deviceId, MqttClientOptionsBuilderTlsParameters tlsOptions, TimeSpan timeOut)
         {
             _brokerUrl = brokerUrl;
             _mqttConnectOptions = new MqttClientOptionsBuilder()
@@ -38,11 +39,12 @@ namespace AstarteDeviceSDKCSharp.Transport.MQTT
             .WithTcpServer(_brokerUrl.Host, _brokerUrl.Port)
             .WithTls(tlsOptions)
             .WithCleanSession(false)
-            .WithCommunicationTimeout(TimeSpan.FromSeconds(60))
             .WithKeepAlivePeriod(TimeSpan.FromSeconds(60))
             .WithSessionExpiryInterval(0)
+            .WithTimeout(timeOut)
             .Build();
 
+            _timeOut = timeOut;
             _clientId = $"{astarteRealm}/{deviceId}";
         }
 
@@ -50,6 +52,8 @@ namespace AstarteDeviceSDKCSharp.Transport.MQTT
 
         public string GetClientId() => _clientId;
 
-        public IMqttClientOptions GetMqttConnectOptions() => _mqttConnectOptions;
+        public TimeSpan GetTimeOut() => _timeOut;
+
+        public MqttClientOptions GetMqttConnectOptions() => _mqttConnectOptions;
     }
 }